+++ deno.json
@@ -5,12 +5,12 @@
     "serve": "supabase functions serve --no-verify-jwt",
     "studio": "open http://localhost:54323",
     "tests": "deno test --allow-all supabase/functions/tests/*",
-    "format": "deno fmt supabase/*",
-    "lint": "deno lint supabase/*",
+    "format": "deno fmt src/shared/* supabase/*",
+    "lint": "deno lint src/shared/* supabase/*",
     "migration": "supabase migration new",
     "migrate": "supabase migration up",
-    "update-types-local": "supabase gen types typescript --local > supabase/functions/_shared/db/database.types.ts",
-    "update-types-linked": "supabase gen types typescript --linked > supabase/functions/_shared/db/database.types.ts"
+    "update-types-local": "supabase gen types typescript --local > src/shared/db.types.ts",
+    "update-types-linked": "supabase gen types typescript --linked > src/shared/db.types.ts"
   },
   "imports": {
     "@std/assert": "jsr:@std/assert@1"
