--- frontend/.env.example
@@ -1,3 +1,0 @@
-VITE_SUPABASE_URL=your_supabase_project_url
-VITE_SUPABASE_ANON_KEY=your_supabase_anon_key
-
