--- frontend/README.md
@@ -1,3 +1,0 @@
-# Frontend
-
-Display some music data from database.
