--- frontend/src/app.tsx
@@ -1,16 +1,0 @@
-import './layout/app.css'
-import './layout/index.css'
-import {ListenedList} from "./tracks/listened-list.tsx";
-
-function App() {
-  return (
-    <>
-      <div className='tracks'>
-      <h2>Last listened</h2>
-        <ListenedList />
-      </div>
-    </>
-  )
-}
-
-export default App
