+++ src/web/.gitignore
@@ -10,10 +10,11 @@ lerna-debug.log*
 node_modules
 dist
 dist-ssr
+.vite
 *.local
 
 # Editor directories and files
-.vscode/*
+.vscode/extensions.json
 !.vscode/extensions.json
 .idea
 .DS_Store
