+++ src/web/README.md
@@ -1,0 +1,19 @@
+# Vite + Deno + React + TypeScript
+
+## Running
+
+You need to have Deno v2.0.0 or later installed to run this repo.
+
+Start a dev server:
+
+```
+$ deno task dev
+```
+
+## Deploy
+
+Build production assets:
+
+```
+$ deno task build
+```
