+++ src/web/vite.config.ts
@@ -1,7 +1,8 @@
 import { defineConfig } from 'vite'
+import deno from '@deno/vite-plugin'
 import react from '@vitejs/plugin-react'
 
 // https://vite.dev/config/
 export default defineConfig({
-  plugins: [react()],
+  plugins: [deno(), react()],
 })
