+++ supabase/functions/_shared/db/db.ts
@@ -1,5 +1,5 @@
 import { SupabaseClient } from "https://esm.sh/@supabase/supabase-js@2";
-import { Database } from "./database.types.ts";
+import { Database } from "../../../../src/shared/database.types.ts";
 
 export interface BaseTableInterface {
   readonly supabase: SupabaseClient | undefined;
